requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
spotipy==2.24.0
waitress==3.0.0